backend/requirements.txt
fastapi
uvicorn
pydantic
ultralytics
requests
opencv-python
pillow
